make chat core implementation internal (#22)
Previously, the class `ChatCore` is tagged `@public` and is exported for the sake of documentation generation.

To resolve this and have the class as truly internal so user will have to use the provider function, this PR creates a public `ChatCore` interface that will be exported. Then the class `ChatCoreImpl` will implement that interface and will be made internal and will not be exported. As such, the documentation is still able to generate without errors.


J=CLIP-411
TEST=manual

see that the generated docs are as expected with the ChatCore interfaces and all internal tags are hidden. See that the class cannot be imported.