chat-core-aws-connect: add resetSession (#39)
Adds the `resetSession` method to ChatCoreAwsConnect, for use
by ChatHeadless when the conversation is reset in the middle of
an agent session.

Also implicitly calls `disconnectParticipant` if the session
is not already undefined.

TEST=manual,auto

Added unit tests in chat-core-aws-connect and chat-headless,
saw expected results. Added browser testing to chat-headless, 
saw no issues when restarting conversation.