chat-core-zendesk: Support Zendesk Agent integration (#45)
Created a new core library, `chat-core-zendesk` to support zendesk integration. This integration relies on [Sunco Web Messaging SDK](https://github.com/zendesk/sunshine-conversations-web) to create a live messaging connection to Zendesk agent platform. The interface follows ChatEventClient and is compatible to use with chat-headless.

For this integration to work, the bot must be configured to use Zendesk integration with a zendesk agent handoff step. And a web integration ID (e.g. zendesk widget id) must be provided to `chat-core-zendesk` to work properly.

Also updated the test site to support both amazon connect and zendesk core library extension. and update the UI to be cleaner.

NOTE: This package is meant to be **INTERNAL** at this moment.

J=CLIP-1516
TEST=manual&auto

see that it works in test site. Also verified it works in chat-ui-react test-site (see video in channel)

https://github.com/user-attachments/assets/184b06f0-6e46-4a31-a2e6-b3989dd2f8dc