chat-core-zendesk: fetch new user's first conversation ID #48
the returned conversation ID when invoking `Smooch.createConversation` for a new user is `TEMPORARY_CONVERSATION`. In order the the rest of the integration logic to work, we need the actual conversation ID. this PR updates it to re-fetch using `Smooch.getDisplayedConversation` to get the actual id for the session. Subsequent conversation creation do return the proper ID so this additional logic is not needed there

J=CLIP-1516
TEST=manual

clear all cookies and values in local/session storage to start as a new user, see that it no longer throws an error on the first conversation.