chat-core-zendesk: handle resolved ticket event #49
unlike deleting ticket, which trigger a passControl event and is handled by the internal bot in the switchboard, agent marking a ticket a resolved will not transfer control. Instead, zendesk will send a csat message (e.g. "how did the chat go?") to the conversation.

This PR will replace that message with `The agent has resolved the ticket. Further assistance will now be provided by the bot.` to inform the user and closes the session. The bot in the frontend will then resume the conversation on subsequent user messages.

J=CLIP-1530
TEST=manual

see that I get the custom message instead of "how did the chat go?" message from zendesk. See that the bot will resume conversation in the next message.

![Screenshot 2024-09-12 at 4 42 05 PM](https://github.com/user-attachments/assets/ff556439-65cf-47c2-9e7d-745d9dc3d5dc)