chat-core-zendesk: add support for ticketFields
Adds support for the new ticketFields param on the ZendeskHandoff
returned from the Chat API. This allows fields specified in the bot
config to autopopulate in arbitrary custom Zendesk fields.

TEST=manual,auto

Ran test site locally, using a bot with a field's ID set in the
ticketFields. Initiated a support session, saw the field populated
with the specified value.

Wrote a unit test validating that the parameters are used and
formatted correctly, saw it pass.