chat-core-zendesk: add support for session reinitialization
Adds support for restoring a Zendesk agent conversation using the
conversation ID. Requires reinitialization of the Zendesk client, which
can only be triggered on the first message sent after the new page is
loaded. This leads to a slight bit of extra lag on that first message,
but otherwise everything seems to work as if the page was never
reloaded.

TEST=manual

Used this locally with upcoming chat-headless change to support this new
method, saw expected results across page loads and within the Zendesk platform.