chat-core-zendesk: reset event listeners on re-initialization (#58)
Previously, once a messaging convo/ticket is closed, we reset the session and unbind all event listeners from zendesk to SDK. On re-initialization for another request to speak to agent, we failed to setup the event listeners again. So any messages sent from agent on the second ticket does NOT go through the SDK and display to user.

This CR ensures that event listeners are cleared and setup on each init call, and the message events should apply for the latest relevant conversation only.

J=CLIP-1664
TEST=manual

started a messaging convo with agent, closed the ticket. Request agent again, messaged as user and agent, see that all messages are displayed as expected on both Zendesk and Chat SDK.