chat-core-zendesk: workaround to inject ticket id to bot message (#60)
This is part of a workaround in order enable the bot message to include the ticket ID when live messaging session has ended.
Zendesk does not have native support to easily access ticket ID within Sunco system and Zendesk system also cannot easily access Sunco conversation ID. So the workaround is as follow:
- SDK will provide the conversation ID in a custom ticket field
- Zendesk will have a trigger upon ticket creation with the agent handoff tag AND the conversation ID field is present, which will invoke a webhook to call a Sunco API to inject the ticket ID into the Sunco conversation's metadata
- When the messaging session is closed. The bot in the switchboard API will grab the ticket in the metadata and include it in its messaging when informing user of the closed session.

This PR implements the first step of the workaround. Additionally, the closing message on CSAT message type is removed as it should be now handled by the bot webhook in the switchboard as well.

Thread context: https://yext.slack.com/archives/C01CSDL5US0/p1733870484402519

J=CLIP-1649
TEST=manual&auto

added unit tests
setup trigger and webhook in the zendesk sandbox account and spin up local test site. Verified that the created ticket for the live messaging session contains the conversation ID in the custom field